Add packet statistics
Add per-SNMP packet statistics.
* Update Go modules to latest.
* Add exporter-wide metrics for sent packets and retries.
* Add per-scrape metrics for packets and retries.

The duplication betweeen the per-scrape and exporter-wide metrics is
intentional, so that packet histogram data can be gathered. This is to
keep the cardinality of the histogram data sane.

+++ config/config.go
@@ -63,7 +63,7 @@ type Config map[string]*Module
 
 type WalkParams struct {
 	Version        int           `yaml:"version,omitempty"`
-	MaxRepetitions uint8         `yaml:"max_repetitions,omitempty"`
+	MaxRepetitions uint32        `yaml:"max_repetitions,omitempty"`
 	Retries        int           `yaml:"retries,omitempty"`
 	Timeout        time.Duration `yaml:"timeout,omitempty"`
 	Auth           Auth          `yaml:"auth,omitempty"`
